+++ Sources/TuistAutomation/Utilities/Formatter.swift
@@ -3,18 +3,26 @@ import TuistCore
 import TuistSupport
 
 protocol Formatting {
-    func buildArguments() throws -> [String]
+    func buildArguments(quiet: Bool) throws -> [String]
 }
 
 final class Formatter: Formatting {
+    private struct Arguments {
+        static let quieter: String = "--quieter"
+    }
     private let binaryLocator: BinaryLocating
 
     init(binaryLocator: BinaryLocating = BinaryLocator()) {
         self.binaryLocator = binaryLocator
     }
 
-    func buildArguments() throws -> [String] {
-        let xcbeautifyPath = try binaryLocator.xcbeautifyPath()
-        return [xcbeautifyPath.pathString]
+    func buildArguments(quiet: Bool) throws -> [String] {
+        var args: [String] = [try binaryLocator.xcbeautifyPath().pathString]
+        
+        if quiet {
+            args.append(Arguments.quieter)
+        }
+        
+        return args
     }
 }